@@ -1,7 +1,7 @@
 from django.forms import ModelChoiceField
 from django_api_forms import Form
 
-from apps.core.models import Project, Remote, Service
+from apps.core.models import Project, Remote
 
 
 class TemporaryUserForms:
