@@ -5,7 +5,6 @@ from uuid import UUID
 from porcupine.base import Serializer
 
 from apps.core.serializers.service import ServiceSerializer
-from apps.core.services.variables import VariablesService
 
 
 class RemoteSerializer:
@@ -30,8 +29,9 @@ class RemoteSerializer:
 
             return serialized_project
 
-        @staticmethod
-        def resolve_variables(data, **kwargs) -> dict:
-            request = kwargs['request']
-
-            return VariablesService.create(request=request, variables=data.variables).get_variables()
+        # @staticmethod
+        # def resolve_variables(data, **kwargs) -> dict:
+        #     request = kwargs['request']
+        #
+        #     return VariablesService.create(request=request, variables=data.variables).get_variables()
+#
